+++ lib/Supplemental.php
@@ -3,8 +3,6 @@
 namespace ICanBoogie\CLDR;
 
 use Closure;
-use ICanBoogie\Accessor\AccessorTrait;
-use ICanBoogie\CLDR\Supplemental\CurrencyData;
 
 /**
  * Representation of the "supplemental" section.
@@ -18,17 +16,9 @@ use ICanBoogie\CLDR\Supplemental\CurrencyData;
  *
  * echo $supplemental['calendarPreferenceData']['001']; // gregorian
  * </pre>
- *
- * @property-read CurrencyData $currency_data
- * @uses self::get_currency_data()
  */
 final class Supplemental extends AbstractSectionCollection implements Warmable
 {
-	/**
-	 * @uses get_currency_data
-	 */
-	use AccessorTrait;
-
 	/**
 	 * Where _key_ is a property, matching a CLDR filename, and _value_ is an array path under "supplemental".
 	 */
@@ -64,14 +54,6 @@ final class Supplemental extends AbstractSectionCollection implements Warmable
 		'windowsZones'           => 'windowsZones',
 
 	];
-
-	private CurrencyData $currency_data;
-
-	private function get_currency_data(): CurrencyData
-	{
-		/* @phpstan-ignore-next-line */
-		return $this->currency_data ??= new CurrencyData($this['currencyData']);
-	}
 
 	public function offsetExists($offset): bool
 	{