+++ tests/LocalizedCurrencyTest.php
@@ -14,34 +14,34 @@ final class LocalizedCurrencyTest extends TestCase
 	use StringHelpers;
 
 	private static Currency $currency;
-	private static LocalizedCurrency $localized;
+	private static LocalizedCurrency $sut;
 
 	public static function setUpBeforeClass(): void
 	{
-		self::$currency = new Currency(get_repository(), 'IEP');
-		self::$localized = new LocalizedCurrency(self::$currency, locale_for('fr'));
+		self::$currency = Currency::of('IEP');
+		self::$sut = new LocalizedCurrency(self::$currency, locale_for('fr'));
 	}
 
 	public function test_name(): void
 	{
-		$this->assertEquals("livre irlandaise", self::$localized->name);
+		$this->assertEquals("livre irlandaise", self::$sut->name);
 	}
 
 	public function test_name_for(): void
 	{
-		$this->assertEquals("livre irlandaise", self::$localized->name);
-		$this->assertEquals("livre irlandaise", self::$localized->name_for(1));
-		$this->assertEquals("livres irlandaises", self::$localized->name_for(10));
+		$this->assertEquals("livre irlandaise", self::$sut->name);
+		$this->assertEquals("livre irlandaise", self::$sut->name_for(1));
+		$this->assertEquals("livres irlandaises", self::$sut->name_for(10));
 	}
 
 	public function test_get_symbol(): void
 	{
-		$this->assertEquals("£IE", self::$localized->symbol);
+		$this->assertEquals("£IE", self::$sut->symbol);
 	}
 
 	public function test_localize(): void
 	{
-		$localized = self::$currency->localize('en');
+		$localized = locale_for('en')->localize(self::$currency);
 		$this->assertInstanceOf(LocalizedCurrency::class, $localized);
 		$this->assertEquals("Irish Pound", $localized->name);
 	}
@@ -49,8 +49,8 @@ final class LocalizedCurrencyTest extends TestCase
 	#[DataProvider('provide_test_format')]
 	public function test_format(string $currency_code, string $locale_id, float|int $number, string $expected): void
 	{
-		$currency = new Currency(get_repository(), $currency_code);
-		$localized = $currency->localize($locale_id);
+		$currency = Currency::of($currency_code);
+		$localized = new LocalizedCurrency($currency, locale_for($locale_id));
 		$this->assertEquals($expected, $localized->format($number));
 	}
 
@@ -78,8 +78,8 @@ final class LocalizedCurrencyTest extends TestCase
 		float $number,
 		string $expected
 	): void {
-		$currency = new Currency(get_repository(), $currency_code);
-		$localized = $currency->localize($locale_id);
+		$currency = Currency::of($currency_code);
+		$localized = new LocalizedCurrency($currency, locale_for($locale_id));
 		$actual = $localized->format($number, LocalizedCurrencyFormatter::PATTERN_ACCOUNTING);
 
 		$this->assertStringSame($expected, $actual);